+++ install.sh
@@ -18,33 +18,21 @@
 #  2: Error
 function log
 {
-  if [[ "$#" -gt 0 ]]; then
-    local -r LOGLEVEL="$1" TEXT="${*:2}" Z='\e[0m'
+  if [[ "$#" -gt 0 ]]; then local -r LOGLEVEL="$1" TEXT="${*:2}" Z='\e[0m'
     if [[ "$LOGLEVEL" =~ [(-2)-2] ]]; then
       case "$LOGLEVEL" in
-        -2)
-           local -r CYAN='\e[1;36m'
-           printf "${CYAN}DEBUG${Z} %s\n" "$TEXT" >&2
+        -2) local -r CYAN='\e[1;36m'; printf "${CYAN}DEBUG${Z} %s\n" "$TEXT" >&2
            ;;
-        -1)
-           local -r BLUE='\e[1;34m'
-           printf "${BLUE}INFO${Z} %s\n" "$TEXT"
+        -1) local -r BLUE='\e[1;34m'; printf "${BLUE}INFO${Z} %s\n" "$TEXT"
            ;;
-         0)
-           local -r GREEN='\e[1;32m'
-           printf "${GREEN}SUCCESS${Z} %s\n" "$TEXT"
+         0) local -r GREEN='\e[1;32m'; printf "${GREEN}SUCCESS${Z} %s\n" "$TEXT"
            ;;
-         1)
-           local -r YELLOW='\e[1;33m'
-           printf "${YELLOW}WARNING${Z} %s\n" "$TEXT"
+         1) local -r YELLOW='\e[1;33m'; printf "${YELLOW}WARNING${Z} %s\n" "$TEXT"
            ;;
-         2)
-           local -r RED='\e[1;31m'
-           printf "${RED}ERROR${Z} %s\n" "$TEXT" >&2
+         2) local -r RED='\e[1;31m'; printf "${RED}ERROR${Z} %s\n" "$TEXT" >&2
            ;;
       esac
-    else
-      log 2 "Invalid log level: [Debug: -2|Info: -1|Success: 0|Warning: 1|Error: 2]"
+    else log 2 "Invalid log level: [Debug: -2|Info: -1|Success: 0|Warning: 1|Error: 2]"
     fi
   fi
 }
@@ -68,10 +56,7 @@ function isRoot
 function isUser
 {
   [[ "$#" -ne 1 ]] && return 2
-  if id -u "$1" &>/dev/null; then
-    return 0
-  else
-    return 1
+  if id -u "$1" &>/dev/null; then return 0; else return 1
   fi
 }
 
@@ -310,16 +295,9 @@ function hasText
 # 2: Missing command argument to check
 function hasCMD
 {
-  if [[ "$#" -eq 1 ]]; then
-    local -r CHECK="$1"
-    if command -v "$CHECK" &>/dev/null; then
-      return 0
-    else
-      return 1
-    fi
-  else
-    return 2
-  fi
+  if [[ "$#" -eq 1 ]]; then local -r CHECK="$1"
+    if command -v "$CHECK" &>/dev/null; then return 0; else return 1; fi
+  else return 2; fi
 }
 
 # Installs package(s) using the package manager and pre-configured options
@@ -331,42 +309,28 @@ function hasCMD
 # 4: Not running as root/sudo
 function installPKG
 {
-  if [[ "$#" -eq 0 ]]; then
-    log 2 "Requires: [PKG(s) to install]"
-    return 3
-  elif [[ "$EUID" -ne 0 ]]; then
-    log 2 "Requires root privileges"
-    return 4 
-  else
-    local -r OPTIONS=(--quiet --assume-yes --no-show-upgraded --auto-remove=true --no-install-recommends)
-    local -r ROOTUPDATE=(apt-get "${OPTIONS[@]}" update) \
-             ROOTINSTALL=(apt-get "${OPTIONS[@]}" install)
-    local PKG=()
-    IFS=' ' read -ra PKG <<<"$@"
-    log -1 "Updating apt lists"
-    if "${ROOTUPDATE[@]}" &>/dev/null; then
-      log 0 "Apt list updated"
-    else
-      log 2 "Couldn't update apt lists"
-      return 1
-    fi
-    log -1 "Installing ${PKG[*]}"
-    if DEBIAN_FRONTEND=noninteractive "${ROOTINSTALL[@]}" "${PKG[@]}"; then
-      log 0 "Installation completed"
-      return 0
-    else
-      log 2 "Something went wrong during installation"
-      return 1
-    fi
+  if [[ "$#" -eq 0 ]]; then log 2 "Requires: [PKG(s) to install]"; return 3
+  elif [[ "$EUID" -ne 0 ]]; then log 2 "Requires root privileges"; return 4 
+  else local -r OPTIONS=(--quiet --assume-yes --no-show-upgraded --auto-remove=true --no-install-recommends)
+       local -r ROOTUPDATE=(apt-get "${OPTIONS[@]}" update) \
+                ROOTINSTALL=(apt-get "${OPTIONS[@]}" install)
+       local PKG=()
+       IFS=' ' read -ra PKG <<<"$@"
+       log -1 "Updating apt lists"
+       if "${ROOTUPDATE[@]}" &>/dev/null; then log 0 "Apt list updated"
+       else log 2 "Couldn't update apt lists"; return 1
+       fi; log -1 "Installing ${PKG[*]}"
+       if DEBIAN_FRONTEND=noninteractive "${ROOTINSTALL[@]}" "${PKG[@]}"; then
+         log 0 "Installation completed"; return 0
+       else log 2 "Something went wrong during installation"; return 1
+       fi
   fi
 }
 
 function add_install_variable
 {
-  declare -x -a INSTALL_VARIABLES
-  INSTALL_VARIABLES+=("$@")
-  if ! hasText 'INSTALL_VARIABLES' "${INSTALL_VARIABLES[@]}"; then
-    add_install_variable INSTALL_VARIABLES
+  declare -x -a INSTALL_VARIABLES; INSTALL_VARIABLES+=("$@")
+  if ! hasText 'INSTALL_VARIABLES' "${INSTALL_VARIABLES[@]}"; then add_install_variable INSTALL_VARIABLES
   fi
 }
 
@@ -377,20 +341,12 @@ function clean_install_variables
 
 function clean_install_tmp
 {
-  if ! isSet TMPDIR; then
-    log 2 "Variable not set: TMPDIR"
-    return 1
-  elif ! cd -; then
-    log 2 "Unable to change directory to: -"
-    return 1
-  elif ! isDirectory "$TMPDIR"; then
-    log 2 "Directory not found: TMPDIR"
-    return 1
+  if ! isSet TMPDIR; then log 2 "Variable not set: TMPDIR"; return 1
+  elif ! cd -; then log 2 "Unable to change directory to: -"; return 1
+  elif ! isDirectory "$TMPDIR"; then log 2 "Directory not found: TMPDIR"; return 1
   else
-    if isRoot; then
-      rm --recursive --force "$TMPDIR"
-    else
-      sudo rm --recursive --force "$TMPDIR"
+    if isRoot; then rm --recursive --force "$TMPDIR"
+    else sudo rm --recursive --force "$TMPDIR"
     fi
   fi
 }
@@ -398,18 +354,13 @@ function clean_install_tmp
 function clean_install_script
 {
   log -1 "Cleaning up from install script"
-  if isSet TMPDIR; then
-    if isDirectory "$TMP"; then
-      clean_install_tmp
-    fi
+  if isSet TMPDIR; then if isDirectory "$TMP"; then clean_install_tmp
+                        fi
   fi
-  if isSet INSTALL_VARIABLES; then
-    clean_install_variables
+  if isSet INSTALL_VARIABLES; then clean_install_variables
   fi
-  if isFile '/.ncp-image'; then
-    rm /.ncp-image
-  fi
-  log 0 "Cleaned up from install script"
+  if isFile '/.ncp-image'; then rm /.ncp-image
+  fi; log 0 "Cleaned up from install script"
 }
 
 
@@ -418,15 +369,11 @@ function clean_install_script
 ########################
 
 
-if ! isRoot; then
-  log 2 "Must be run as root or with sudo, try: 'sudo ./${BASH_SOURCE[0]##*/}'"
-  exit 1
+if ! isRoot; then log 2 "Must be run as root or with sudo, try: 'sudo ./${BASH_SOURCE[0]##*/}'"; exit 1
 fi
 
-if isSet DBG && notZero "$DBG"; then
-  set -e"$DBG"
-else
-  set -e
+if isSet DBG && notZero "$DBG"; then set -e"$DBG"
+else set -e
 fi
 
 # Repository owner
@@ -480,8 +427,7 @@ fi
 export PATH
 
 # Check for existing MariaDB/MySQL install
-if hasCMD mysqld; then
-  log 1 "Existing MySQL configuration will be changed"
+if hasCMD mysqld; then log 1 "Existing MySQL configuration will be changed"
   if isSet DBNAME; then if mysql -e 'use '"$DBNAME"'' &>/dev/null; then { log 2 "Database exists: $DBNAME"; exit 1; }; fi
   else if mysql -e 'use nextcloud' &>/dev/null; then log 2 "Database exists: nextcloud"; exit 1; fi
   fi
@@ -533,9 +479,7 @@ fi
 if ! touch '/.ncp-image'; then log 2 "Failed to create file: /.ncp-image"; exit 1; fi
 
 # Create the local NextcloudPi configuration directory
-if ! mkdir --parents '/usr/local/etc/ncp-config.d'; then
-  log 2 "Failed creating directory: /usr/local/etc/ncp-config.d"
-  exit 1
+if ! mkdir --parents '/usr/local/etc/ncp-config.d'; then log 2 "Failed creating directory: /usr/local/etc/ncp-config.d"; exit 1
 fi
 
 # Check so the local & build configuration directories exists and
@@ -544,26 +488,17 @@ if isDirectory 'etc/ncp-config.d'; then
   if isDirectory '/usr/local/etc/ncp-config.d'; then
     if isFile 'etc/ncp-config.d/nc-nextcloud.cfg'; then
       if ! cp 'etc/ncp-config.d/nc-nextcloud.cfg' '/usr/local/etc/ncp-config.d/nc-nextcloud.cfg'; then
-        log 2 "Failed to copy file: nc-nextcloud.cfg"
-        exit 1
+        log 2 "Failed to copy file: nc-nextcloud.cfg"; exit 1
       fi
-    else
-      log 2 "File not found: etc/ncp-config.d/nc-nextcloud.cfg"
-      exit 1
+    else log 2 "File not found: etc/ncp-config.d/nc-nextcloud.cfg"; exit 1
     fi
-  else
-    log 2 "Directory not found: /usr/local/etc/ncp-config.d"
-    exit 1
+  else log 2 "Directory not found: /usr/local/etc/ncp-config.d"; exit 1
   fi
-else
-  log 2 "Directory not found: etc/ncp-config.d"
-  exit 1
+else log 2 "Directory not found: etc/ncp-config.d"; exit 1
 fi
 
 if isFile "$LIBRARY"; then
-  if ! cp "$LIBRARY" '/usr/local/etc/library.sh'; then
-    log 2 "Failed to copy file: $LIBRARY"
-    exit 1
+  if ! cp "$LIBRARY" '/usr/local/etc/library.sh'; then log 2 "Failed to copy file: $LIBRARY"; exit 1
   fi
   LIBRARY='/usr/local/etc/library.sh'
   declare -x -g LIBRARY
@@ -571,9 +506,7 @@ if isFile "$LIBRARY"; then
 fi
 
 if isFile "$NCPCFG"; then
-  if ! cp "$NCPCFG" '/usr/local/etc/ncp.cfg'; then
-    log 2 "Failed to copy file: ncp.cfg $NCPCFG"
-    exit 1
+  if ! cp "$NCPCFG" '/usr/local/etc/ncp.cfg'; then log 2 "Failed to copy file: ncp.cfg $NCPCFG"; exit 1
   fi
   NCPCFG='/usr/local/etc/ncp.cfg'
   declare -x -g NCPCFG
@@ -581,64 +514,42 @@ if isFile "$NCPCFG"; then
 fi
 
 if isDirectory "$NCP_TEMPLATES_DIR"; then
-  if ! cp -r "$NCP_TEMPLATES_DIR" '/usr/local/etc/'; then
-    log 2 "Failed to copy templates: $NCP_TEMPLATES_DIR"
-    exit 1
+  if ! cp -r "$NCP_TEMPLATES_DIR" '/usr/local/etc/'; then log 2 "Failed to copy templates: $NCP_TEMPLATES_DIR"; exit 1
   fi
   NCP_TEMPLATES_DIR='/usr/local/etc/ncp-templates'
-  declare -x -g NCP_TEMPLATES_DIR
-  log -2 "NCP_TEMPLATES_DIR: $NCP_TEMPLATES_DIR"
-else
-  log 2 "Directory not found: $NCP_TEMPLATES_DIR"
-  exit 1
+  declare -x -g NCP_TEMPLATES_DIR; log -2 "NCP_TEMPLATES_DIR: $NCP_TEMPLATES_DIR"
+else log 2 "Directory not found: $NCP_TEMPLATES_DIR"; exit 1
 fi
 
 # cp etc/library.sh /usr/local/etc/
 # cp etc/ncp.cfg /usr/local/etc/
 # cp -r etc/ncp-templates /usr/local/etc/
 
-if isFile 'lamp.sh'; then
-  install_app lamp.sh
-else
-  log 2 "File not found: lamp.sh"
-  exit 1
+if isFile 'lamp.sh'; then install_app lamp.sh
+else log 2 "File not found: lamp.sh"; exit 1
 fi
 
-if isFile 'bin/ncp/CONFIG/nc-nextcloud.sh'; then
-  install_app bin/ncp/CONFIG/nc-nextcloud.sh
-else
-  log 2 "File not found: bin/ncp/CONFIG/nc-nextcloud.sh"
-  exit 1
+if isFile 'bin/ncp/CONFIG/nc-nextcloud.sh'; then install_app bin/ncp/CONFIG/nc-nextcloud.sh
+else log 2 "File not found: bin/ncp/CONFIG/nc-nextcloud.sh"; exit 1
 fi
 
-if isFile 'bin/ncp/CONFIG/nc-nextcloud.sh'; then
-  run_app_unsafe bin/ncp/CONFIG/nc-nextcloud.sh
-else
-  log 2 "File not found: bin/ncp/CONFIG/nc-nextcloud.sh"
-  exit 1
+if isFile 'bin/ncp/CONFIG/nc-nextcloud.sh'; then run_app_unsafe bin/ncp/CONFIG/nc-nextcloud.sh
+else log 2 "File not found: bin/ncp/CONFIG/nc-nextcloud.sh"; exit 1
 fi
 
-if isFile '/usr/local/etc/ncp-config.d/nc-nextcloud.cfg'; then
-  rm /usr/local/etc/ncp-config.d/nc-nextcloud.cfg    # armbian overlay is ro
-else
-  log 2 "File not found: /usr/local/etc/ncp-config.d/nc-nextcloud.cfg"
-  exit 1
+# armbian overlay is ro
+if isFile '/usr/local/etc/ncp-config.d/nc-nextcloud.cfg'; then rm /usr/local/etc/ncp-config.d/nc-nextcloud.cfg
+else log 2 "File not found: /usr/local/etc/ncp-config.d/nc-nextcloud.cfg"; exit 1
 fi
 
 systemctl restart mysqld # TODO this shouldn't be necessary, but somehow it's needed in Debian 9.6. Fixme
 
-if isFile 'ncp.sh'; then
-  install_app ncp.sh
-else
-  log 2 "File not found: ncp.sh"
-  exit 1
+if isFile 'ncp.sh'; then install_app ncp.sh
+else log 2 "File not found: ncp.sh"; exit 1
 fi
 
-if isFile 'bin/ncp/CONFIG/nc-init.sh'; then
-  run_app_unsafe bin/ncp/CONFIG/nc-init.sh
-else
-  log 2 "File not found: bin/ncp/CONFIG/nc-init.sh"
-  exit 1
+if isFile 'bin/ncp/CONFIG/nc-init.sh'; then run_app_unsafe bin/ncp/CONFIG/nc-init.sh
+else log 2 "File not found: bin/ncp/CONFIG/nc-init.sh"; exit 1
 fi
 
 log -1 "Moving data directory to: /opt/ncdata"
@@ -651,52 +562,37 @@ if ! isFile "/usr/local/etc/ncp-config.d/nc-datadir.cfg" ]]; then
     if ! cp 'etc/ncp-config.d/nc-datadir.cfg' '/usr/local/etc/ncp-config.d/nc-datadir.cfg'; then
       log 2 "Filed to copy file: etc/ncp-config.d/nc-datadir.cfg | To: /usr/local/etc/ncp-config.d/nc-datadir.cfg"
     fi
-  else
-    log 2 "File not found: etc/ncp-config.d/nc-datadir.cfg"
-    exit 1
+  else log 2 "File not found: etc/ncp-config.d/nc-datadir.cfg"; exit 1
   fi
 fi
 
 if isFile 'bin/ncp/CONFIG/nc-datadir.sh'; then
   DISABLE_FS_CHECK=1 NCPCFG="/usr/local/etc/ncp.cfg" run_app_unsafe 'bin/ncp/CONFIG/nc-datadir.sh'
-else
-  log 2 "File not found: bin/ncp/CONFIG/nc-datadir.sh"
-  exit 1
+else log 2 "File not found: bin/ncp/CONFIG/nc-datadir.sh"; exit 1
 fi
 
 if notZero "$should_rm_datadir_cfg"; then
-  if isFile '/usr/local/etc/ncp-config.d/nc-datadir.cfg'; then
-    rm '/usr/local/etc/ncp-config.d/nc-datadir.cfg'
-  else
-    log 2 "File not found: /usr/local/etc/ncp-config.d/nc-datadir.cfg"
-    exit 1
+  if isFile '/usr/local/etc/ncp-config.d/nc-datadir.cfg'; then rm '/usr/local/etc/ncp-config.d/nc-datadir.cfg'
+  else log 2 "File not found: /usr/local/etc/ncp-config.d/nc-datadir.cfg"; exit 1
   fi
 fi
 
-if isFile '/.ncp-image'; then
-  rm '/.ncp-image'
+if isFile '/.ncp-image'; then rm '/.ncp-image'
 fi
 
 # Skip on Armbian / Vagrant / LXD
 if notZero "$CODE_DIR"; then
   if isFile '/usr/local/bin/ncp-provisioning.sh'; then
     bash /usr/local/bin/ncp-provisioning.sh
-  else
-    log 2 "File not found: /usr/local/bin/ncp-provisioning.sh"
-    exit 1
+  else log 2 "File not found: /usr/local/bin/ncp-provisioning.sh"; exit 1
   fi
 fi
 
-if ! cd -; then
-  log 2 "Failed to change directory to: -"
-  exit 1
+if ! cd -; then log 2 "Failed to change directory to: -"; exit 1
 fi
 
-if isDirectory "$TMPDIR"; then
-  rm --recursive --force "$TMPDIR"
-else
-  log 2 "Directory not found: $TMPDIR"
-  exit 1
+if isDirectory "$TMPDIR"; then rm --recursive --force "$TMPDIR"
+else log 2 "Directory not found: $TMPDIR"; exit 1
 fi
 
 trap - EXIT SIGHUP SIGILL SIGABRT SIGINT
